+++ test-d/iterable-element.ts
@@ -1,4 +1,4 @@
-import {expectType} from 'tsd';
+import {expectAssignable, expectType} from 'tsd';
 import type {IterableElement} from '../index';
 
 declare const iterableElement: IterableElement<ReturnType<typeof secretGenerator>>;
@@ -16,3 +16,25 @@ async function * secretGeneratorAsync() {
 	yield true;
 	yield new Date();
 }
+
+const fruits = new Set(['🍎', '🍌', '🍉'] as const);
+
+type Fruit = IterableElement<typeof fruits>;
+
+expectAssignable<Fruit>('🍎');
+expectAssignable<Fruit>('🍌');
+expectAssignable<Fruit>('🍉');
+
+type VegetableSet = Set<'🥦' | '🥕' | '🌶'>;
+type Vegetable = IterableElement<VegetableSet>;
+
+expectAssignable<Vegetable>('🥦');
+expectAssignable<Vegetable>('🥕');
+expectAssignable<Vegetable>('🌶');
+
+type UserRolesSet = ReadonlySet<'regular' | 'contributor' | 'maintainer'>;
+type UserRole = IterableElement<UserRolesSet>;
+
+expectAssignable<UserRole>('regular');
+expectAssignable<UserRole>('contributor');
+expectAssignable<UserRole>('maintainer');
